# server/swagger.yaml
openapi: 3.0.3
info:
  title: API Login Portfolio
  version: 1.0.0
paths:
  /register:
    post:
      summary: Registra um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
                - password
              properties:
                email:
                  type: string
                  example: usuario@email.com
                username:
                  type: string
                  example: usuario123
                password:
                  type: string
                  example: senha123
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Usuário criado com sucesso
        '400':
          description: Erro nos dados enviados ou usuário já existe
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Usuário já existe

  /login:
    post:
      summary: Faz login do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                email:
                  type: string
                  example: usuario@email.com
                username:
                  type: string
                  example: luisbatista
                password:
                  type: string
                  example: senha123
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: fake-jwt-token
                  user:
                    type: object
                    properties:
                      username:
                        type: string
                      email:
                        type: string
        '401':
          description: Credenciais inválidas